+++ finality-aleph/Cargo.toml
@@ -18,6 +18,7 @@ network-clique = { package = "network-clique", path = "../clique" }
 aleph-primitives = { package = "primitives", path = "../primitives" }
 legacy-aleph-aggregator = { package = "aggregator", git = "https://github.com/Cardinal-Cryptography/aleph-node.git", tag = "aggregator-v0.2.1" }
 current-aleph-aggregator = { path = "../aggregator", package = "aggregator" }
+rate-limiter = { package = "rate-limiter", path = "../rate-limiter" }
 
 async-trait = "0.1"
 bytes = "1.0"
