+++ finality-aleph/src/network/tcp.rs
@@ -5,9 +5,9 @@ use codec::{Decode, Encode};
 use derive_more::{AsRef, Display};
 use log::info;
 use network_clique::{
-    rate_limiter::TokenBucket, Dialer, Listener, PeerId, PublicKey, RateLimitingDialer,
-    RateLimitingListener, SecretKey,
+    Dialer, Listener, PeerId, PublicKey, RateLimitingDialer, RateLimitingListener, SecretKey,
 };
+use rate_limiter::TokenBucket;
 use sp_core::crypto::KeyTypeId;
 use tokio::net::{TcpListener, TcpStream, ToSocketAddrs};
