+++ finality-aleph/src/nodes/validator_node.rs
@@ -4,7 +4,8 @@ use aleph_primitives::BlockNumber;
 use bip39::{Language, Mnemonic, MnemonicType};
 use futures::channel::oneshot;
 use log::{debug, error};
-use network_clique::{rate_limiter::TokenBucket, Service, SpawnHandleT};
+use network_clique::{Service, SpawnHandleT};
+use rate_limiter::TokenBucket;
 use sc_client_api::Backend;
 use sc_network_common::ExHashT;
 use sp_consensus::SelectChain;
